requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
email-validator==2.0.0.post2
PyJWT==2.8.0
pytest==7.4.2
Werkzeug==2.3.7
Flask-Migrate==4.0.5
Flask-Caching==2.0.2
Pillow==10.0.0

# Development and testing
flake8==6.1.0
mypy==1.5.1
bandit==1.7.5
pytest-cov==4.1.0
